Changed into a poetry project
Added pyproject.toml and lock.
Partially factored out the code that
interacts with the databse.
Testcase is not main, but needs to
be called from the testing infrastructure.
