+++ .gitignore
@@ -1,0 +1,8 @@
+__pycache__/
+*~
+
+env/
+dist/
+
+data/
+exports.sh
