+++ live_data_analysis/__init__.py
@@ -1,0 +1,4 @@
+from live_data_analysis.data_and_connection import temp_sensors, pressure, solar
+
+TESTDATAPATH="./live_data_analysis/test/"
+from live_data_analysis.test_live_data import tester
