+++ navgreen_base/__init__.py
@@ -1,0 +1,5 @@
+import warnings
+from influxdb_client.client.warnings import MissingPivotFunction
+warnings.simplefilter("ignore", MissingPivotFunction)
+
+from navgreen_base.influx import make_point, write_data, read_data, delete_data
