+++ pyproject.toml
@@ -1,0 +1,22 @@
+[tool.poetry]
+name = "navgreen"
+packages = [ {include = "live_data_analysis"},
+             {include = "hist_data_analysis"} ]
+version = "0.1"
+description = ""
+authors = [ "Tatiana Boura <tat.boura@gmail.com>",
+	    "Stasinos Konstantopoulos <konstant@iit.demokritos.gr>" ]
+readme = "README.md"
+
+[tool.poetry.dependencies]
+python = ">=3.11,<3.13"
+pandas = "^2.1.2"
+influxdb-client = "^1.38.0"
+pymodbus = "^3.5.4"
+
+[build-system]
+requires = ["poetry-core"]
+build-backend = "poetry.core.masonry.api"
+
+[tool.poetry.scripts]
+tester = "live_data_analysis:tester"
